backend/requirements.txt
# Web framework
fastapi==0.115.*
uvicorn[standard]==0.34.*

# Database
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*

# Auth
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt==4.1.*

# Settings
pydantic-settings==2.7.*

# Utilities
python-dateutil==2.9.*
python-multipart==0.0.9

# Testing
pytest==8.3.*
pytest-asyncio==0.24.*
httpx==0.28.*
